Factor out milestoneGroup computation and itemTypeKey
