+++ client/app/bundles/course/lesson_plan/components/LessonPlanGroup.jsx
@@ -8,10 +8,9 @@ import LessonPlanMilestone from '../components/LessonPlanMilestone';
 const propTypes = {
   milestone: PropTypes.instanceOf(Immutable.Map).isRequired,
   items: PropTypes.instanceOf(Immutable.List).isRequired,
-  lessonPlanItemTypeKey: PropTypes.func.isRequired,
 };
 
-const MilestoneGroup = ({ milestone, items, lessonPlanItemTypeKey }) => {
+const MilestoneGroup = ({ milestone, items }) => {
   const componentKey = node => node.get('lesson_plan_element_class') + node.get('id');
   return (
     <Element name={`milestone-group-${milestone.get('id')}`}>
@@ -21,7 +20,7 @@ const MilestoneGroup = ({ milestone, items, lessonPlanItemTypeKey }) => {
           items.map(item =>
             <LessonPlanItem
               key={componentKey(item)}
-              {...{ lessonPlanItemTypeKey, item }}
+              {...{ item }}
             />
           )
         }
