+++ client/app/bundles/course/lesson_plan/containers/LessonPlanContainer.jsx
@@ -2,29 +2,108 @@ import React, { PropTypes } from 'react';
 import Immutable from 'immutable';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
+import { injectIntl, defineMessages } from 'react-intl';
 import LessonPlan from '../components/LessonPlan';
 import * as actionCreators from '../actions';
+import lessonPlanItemTypeKey from '../utils';
+
+const translations = defineMessages({
+  priorItemsMilestoneTitle: {
+    id: 'course.lessonPlan.lessonPlan.priorItemsMilestoneTitle',
+    defaultMessage: 'Prior Items',
+    description: 'Title for prior items milestones',
+  },
+});
+
+const propTypes = {
+  milestones: PropTypes.instanceOf(Immutable.List).isRequired,
+  items: PropTypes.instanceOf(Immutable.List).isRequired,
+  hiddenItemTypes: PropTypes.instanceOf(Immutable.List).isRequired,
+  dispatch: PropTypes.func.isRequired,
+  intl: PropTypes.shape({
+    formatMessage: PropTypes.func.isRequired,
+  }).isRequired,
+};
 
 function mapStateToProps(state) {
   return state.get('lessonPlan').toObject();
 }
 
-const propTypes = {
-  milestones: React.PropTypes.instanceOf(Immutable.List).isRequired,
-  items: React.PropTypes.instanceOf(Immutable.List).isRequired,
-  hiddenItemTypes: React.PropTypes.instanceOf(Immutable.List).isRequired,
-  dispatch: PropTypes.func.isRequired,
-};
+class LessonPlanContainer extends React.Component {
+  /**
+   * Groups lesson plan items with their respective milestones into 'milestone groups'
+   * and sorts items according to their types within each milestone group.
+   */
+  milestoneGroups() {
+    const { milestones, items, hiddenItemTypes, intl } = this.props;
 
-const LessonPlanContainer = (props) => {
-  const { dispatch, milestones, items, hiddenItemTypes } = props;
-  const { toggleItemTypeVisibility } = bindActionCreators(actionCreators, dispatch);
+    const MILESTONE_CLASS_NAME = 'Course::LessonPlan::Milestone';
+    const isMilestone = elem => elem.get('lesson_plan_element_class') === MILESTONE_CLASS_NAME;
 
-  return (
-    <LessonPlan {...{ milestones, items, hiddenItemTypes, toggleItemTypeVisibility }} />
-  );
-};
+    // Filter away items that the user has hidden
+    const visibleItems = items.filter((item) => {
+      const type = lessonPlanItemTypeKey(item.get('lesson_plan_item_type'));
+      return !hiddenItemTypes.includes(type);
+    });
+
+    // Combine milestones and visible items, then order them chronologically
+    const sortedLessonPlanElements = milestones.concat(visibleItems)
+      .map(elem => elem.merge({ start_at: Date.parse(elem.get('start_at')) }))
+      .sortBy(elem => elem.get('start_at'));
+
+    // Create a default milestone to hold items that precede the first milestone,
+    // if one is needed.
+    const seedGroup =
+      sortedLessonPlanElements.isEmpty() || isMilestone(sortedLessonPlanElements.first()) ?
+      [] :
+      [{
+        milestone: Immutable.fromJS({
+          id: 'prior',
+          title: intl.formatMessage(translations.priorItemsMilestoneTitle),
+          start_at: sortedLessonPlanElements.first().get('start_at'),
+        }),
+        items: [],
+      }];
+
+    // Reducer method that takes a flat array of milestones and items,
+    // and chunks them into milestone groups.
+    const groupItemsWithMilestones = (result, elem) => {
+      if (isMilestone(elem)) {
+        result.push({ milestone: elem, items: [] });
+      } else {
+        result[result.length - 1].items.push(elem);
+      }
+      return result;
+    };
+
+    // Maps a milestone group with unsorted items to a duplicate group that has the items sorted.
+    const sortItemsByType = (group) => {
+      const { milestone } = group;
+      const sortedItems =
+        Immutable.fromJS(group.items).sortBy(item => item.get('lesson_plan_item_type'));
+      return { milestone, items: sortedItems };
+    };
+
+    // Build the sorted milestone groups
+    const groups = sortedLessonPlanElements
+      .reduce(groupItemsWithMilestones, seedGroup)
+      .map(sortItemsByType);
+
+    return groups;
+  }
+
+  render() {
+    const actions = bindActionCreators(actionCreators, this.props.dispatch);
+    const childProps = {
+      ...this.props,
+      ...actions,
+      milestoneGroups: this.milestoneGroups(),
+    };
+
+    return (<LessonPlan {...childProps} />);
+  }
+}
 
 LessonPlanContainer.propTypes = propTypes;
 
-export default connect(mapStateToProps)(LessonPlanContainer);
+export default connect(mapStateToProps)(injectIntl(LessonPlanContainer));
